+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Bump Sentry to v6.1.1
 - Save search query to overdue search history when search query is valid
 - Add layout for overdue search screen
+- Handle user interactions for overdue search results
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291